+++ test/Vitorm.Sqlite.MsTest/CommonTest/Orm_Extensions_ExecuteUpdateAsync_Test.cs
@@ -22,7 +22,7 @@ namespace Vitorm.MsTest.CommonTest
 
                 Assert.AreEqual(6, count);
 
-                var userList = userQuery.ToList();
+                var userList = userQuery.OrderBy(m => m.id).ToList();
                 Assert.AreEqual("u_1_4_6", userList.First().name);
                 Assert.AreEqual(DateTime.Parse("2021-01-11 00:00:00"), userList.First().birth);
                 Assert.AreEqual("u_6__", userList.Last().name);
@@ -45,7 +45,7 @@ namespace Vitorm.MsTest.CommonTest
                 Assert.AreEqual(6, count);
 
 
-                var userList = userQuery.ToList();
+                var userList = userQuery.OrderBy(m => m.id).ToList();
                 Assert.AreEqual("u2_1_4_6", userList.First().name);
                 Assert.AreEqual("u2_6__", userList.Last().name);
             }
@@ -68,7 +68,7 @@ namespace Vitorm.MsTest.CommonTest
                 Assert.AreEqual(3, count);
 
 
-                var userList = userQuery.ToList();
+                var userList = userQuery.OrderBy(m => m.id).ToList();
                 Assert.AreEqual("u3_1_4_6", userList[0].name);
                 Assert.AreEqual("u3_3_5_6", userList[2].name);
                 Assert.AreEqual("u2_4__", userList[3].name);