+++ doc/ReleaseNotes.md
@@ -1,6 +1,11 @@
 ﻿# Vitorm ReleaseNotes
 
 -----------------------
+# 2.5.0
+- [Vitorm] remove EntityPropertyReader.cs(Not useful)
+- [Vitorm] add orderBy to test case
+
+-----------------------
 # 2.4.0
 - [Vitorm] DbContext, new method GetEntityDescriptorFromCache
 - [Vitorm] move BeginTransaction to interface and refactor Transaction